backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
sqlalchemy[asyncio]==2.0.28
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
reportlab==4.1.0
python-docx==1.1.0
boto3==1.34.60
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
aiosqlite==0.20.0
greenlet==3.1.1
aiosmtplib>=2.0.0
